+++ tests/testthat/test-ensure.R
@@ -5,10 +5,9 @@ context("ensure")
 #' @param length numeric. The length of the random string to generate.
 #' @param alphabet character. A list of characters to draw from to create the string.
 random_string <- ensure(
-  pre = list(length %is% numeric,
+  pre = list(length %is% numeric, length(length) == 1, length > 0,
     alphabet %is% list || alphabet %is% vector,
-    alphabet %contains_only% simple_string,
-    length > 0),
+    alphabet %contains_only% simple_string),
   post = list(result %is% simple_string, nchar(result) == length),
   function(length, alphabet) {
     paste0(sample(alphabet, length, replace = TRUE), collapse = "")
@@ -46,7 +45,7 @@ describe("precondition validations", {
   })
   test_that("can have multiple errors", {
     expect_error(random_string(-10, list(1, 2, 3)),
-      "Error on alphabet %contains_only% simple_string, length > 0")
+      "Error on length > 0, alphabet %contains_only% simple_string")
   })
   test_that("default args are also checked", {
     add_default <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,