+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Validating a function twice creates an error.
 * Printing a validated function will show the preconditions, postconditions, and the original function, rather than the metaprogramming behind the scenes.
+* Fixed a bug when calling functions with default arguments.
 
 #### v0.0.0.9002