+++ NEWS.md
@@ -3,6 +3,7 @@
 * Validating a function twice creates an error.
 * Printing a validated function will show the preconditions, postconditions, and the original function, rather than the metaprogramming behind the scenes.
 * Fixed a bug when calling functions with default arguments.
+* Clarified `random_string` preconditions to not allow length >1 `length` arguments.
 
 #### v0.0.0.9002